+++ tomcat/src/main/java/com/spring/http/cookie/HttpCookies.java
@@ -1,4 +1,4 @@
-package org.apache.catalina.domain.cookie;
+package com.spring.http.cookie;
 
 import java.util.ArrayList;
 import java.util.List;
