+++ tomcat/src/main/java/com/techcourse/controller/LoginController.java
@@ -9,8 +9,8 @@ import com.spring.controller.Controller;
 import java.io.IOException;
 import java.util.Map;
 import org.apache.catalina.domain.Session;
-import org.apache.catalina.domain.request.HttpRequest;
-import org.apache.catalina.domain.response.HttpResponse;
+import com.spring.http.request.HttpRequest;
+import com.spring.http.response.HttpResponse;
 import org.apache.catalina.manager.SessionManager;
 import org.apache.catalina.util.FileParser;
 import org.slf4j.Logger;
