+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -7,8 +7,8 @@ import com.techcourse.model.User;
 import com.spring.controller.Controller;
 import java.io.IOException;
 import java.util.Map;
-import org.apache.catalina.domain.request.HttpRequest;
-import org.apache.catalina.domain.response.HttpResponse;
+import com.spring.http.request.HttpRequest;
+import com.spring.http.response.HttpResponse;
 import org.apache.catalina.util.FileParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
