+++ tomcat/src/main/java/org/apache/catalina/servlet/HttpServletContainer.java
@@ -7,8 +7,8 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
-import org.apache.catalina.domain.request.HttpRequest;
-import org.apache.catalina.domain.response.HttpResponse;
+import com.spring.http.request.HttpRequest;
+import com.spring.http.response.HttpResponse;
 import org.apache.catalina.servlet.impl.DefaultServlet;
 import org.apache.coyote.util.ResponseUtil;
 import org.slf4j.Logger;
