+++ tomcat/src/main/java/org/apache/catalina/servlet/impl/DefaultServlet.java
@@ -3,8 +3,8 @@ package org.apache.catalina.servlet.impl;
 import com.spring.http.enums.HttpStatus;
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import org.apache.catalina.domain.request.HttpRequest;
-import org.apache.catalina.domain.response.HttpResponse;
+import com.spring.http.request.HttpRequest;
+import com.spring.http.response.HttpResponse;
 import org.apache.catalina.servlet.HttpServlet;
 import org.apache.catalina.util.FileParser;
 import org.slf4j.Logger;
