+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -7,8 +7,8 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.Socket;
 import java.nio.charset.StandardCharsets;
-import org.apache.catalina.domain.request.HttpRequest;
-import org.apache.catalina.domain.response.HttpResponse;
+import com.spring.http.request.HttpRequest;
+import com.spring.http.response.HttpResponse;
 import org.apache.catalina.servlet.HttpServletContainer;
 import org.apache.coyote.Processor;
 import org.apache.coyote.util.HttpRequestParser;
