+++ tomcat/src/main/java/org/apache/coyote/util/ResponseUtil.java
@@ -3,8 +3,8 @@ package org.apache.coyote.util;
 import com.spring.http.enums.HttpStatus;
 import java.io.IOException;
 import org.apache.catalina.connector.ResponseHeaderUtil;
-import org.apache.catalina.domain.request.HttpRequest;
-import org.apache.catalina.domain.response.HttpResponse;
+import com.spring.http.request.HttpRequest;
+import com.spring.http.response.HttpResponse;
 import org.apache.catalina.util.FileParser;
 
 public final class ResponseUtil {
