+++ tomcat/src/test/java/com/techcourse/controller/LoginControllerTest.java
@@ -15,11 +15,11 @@ import com.spring.controller.Controller;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
-import org.apache.catalina.domain.HttpHeader;
-import org.apache.catalina.domain.request.HttpRequest;
-import org.apache.catalina.domain.request.HttpRequestBody;
-import org.apache.catalina.domain.request.RequestStartLine;
-import org.apache.catalina.domain.response.HttpResponse;
+import com.spring.http.HttpHeader;
+import com.spring.http.request.HttpRequest;
+import com.spring.http.request.HttpRequestBody;
+import com.spring.http.request.RequestStartLine;
+import com.spring.http.response.HttpResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
