+++ tomcat/src/test/java/org/apache/catalina/domain/HttpCookieTest.java
@@ -2,7 +2,7 @@ package org.apache.catalina.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.apache.catalina.domain.cookie.HttpCookie;
+import com.spring.http.cookie.HttpCookie;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
