+++ tomcat/src/test/java/org/apache/catalina/domain/HttpCookiesTest.java
@@ -2,8 +2,8 @@ package org.apache.catalina.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.apache.catalina.domain.cookie.HttpCookie;
-import org.apache.catalina.domain.cookie.HttpCookies;
+import com.spring.http.cookie.HttpCookie;
+import com.spring.http.cookie.HttpCookies;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
