+++ tomcat/src/test/java/org/apache/catalina/domain/HttpHeaderTest.java
@@ -2,9 +2,10 @@ package org.apache.catalina.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.spring.http.HttpHeader;
 import java.util.Arrays;
 import java.util.List;
-import org.apache.catalina.domain.cookie.HttpCookies;
+import com.spring.http.cookie.HttpCookies;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
